fastapi_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
torch==2.4.1+cpu
torchvision==0.19.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
google-genai==0.2.0
